requirements.txt
autogen-agentchat==0.5.7
autogen-ext[openai]
pydantic
python-dotenv
fastapi
uvicorn[standard]
-e .
